[#8 (30)] Destructive operations in Lexeme#update
Original behaviour: nothing happened (things that were supposed to be deleted weren't even deleted)
New behaviour: destroy if field and all its 'children' are blank.

Also if one tries to delete a headword form and it is not empty, its form will be changed to its first phonetic form.  If one tries to delete an inflection line, it will replace it with a list of headwords.